+++ OTAnalytics/plugin_ui/base_application.py
